data-ingestion-tool/requirements.txt
flask==2.3.2
clickhouse-driver==0.2.6
pandas==2.0.3
python-dotenv==1.0.0
flask-cors==4.0.0